soap consumer moved from src files to test files
