+++ src/test/java/it/com/proxiad/hangmangame/logic/ranking/soapconsumer/GetTop10ForLastMonthResponse.java
@@ -1,5 +1,5 @@
 
-package com.proxiad.hangmangame.soap.consumer;
+package it.com.proxiad.hangmangame.logic.ranking.soapconsumer;
 
 import java.util.ArrayList;
 import java.util.List;
