+++ src/test/java/it/com/proxiad/hangmangame/logic/ranking/soapconsumer/ObjectFactory.java
@@ -1,5 +1,5 @@
 
-package com.proxiad.hangmangame.soap.consumer;
+package it.com.proxiad.hangmangame.logic.ranking.soapconsumer;
 
 import javax.xml.bind.JAXBElement;
 import javax.xml.bind.annotation.XmlElementDecl;
@@ -10,7 +10,7 @@ import javax.xml.namespace.QName;
 /**
  * This object contains factory methods for each 
  * Java content interface and Java element interface 
- * generated in the com.proxiad.hangmangame.soap.consumer package. 
+ * generated in the it.com.proxiad.hangmangame.logic.ranking.soapconsumer package. 
  * <p>An ObjectFactory allows you to programatically 
  * construct new instances of the Java representation 
  * for XML content. The Java representation of XML 
@@ -30,7 +30,7 @@ public class ObjectFactory {
     private final static QName _GetTop10ForLastMonthResponse_QNAME = new QName("http://ranking.logic.hangmangame.proxiad.com/", "getTop10ForLastMonthResponse");
 
     /**
-     * Create a new ObjectFactory that can be used to create new instances of schema derived classes for package: com.proxiad.hangmangame.soap.consumer
+     * Create a new ObjectFactory that can be used to create new instances of schema derived classes for package: it.com.proxiad.hangmangame.logic.ranking.soapconsumer
      * 
      */
     public ObjectFactory() {
